python-files/requirements.txt
Babel==2.3.4
blinker==1.4
click==6.6
coverage==4.2
decorator==4.0.10
defusedxml==0.4.1
Flask==0.11.1
Flask-Babel==0.11.1
Flask-HTTPAuth==3.2.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.13.1
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
passlib==1.6.5
pbr==1.10.0
psycopg2==2.6.2
pycrypto==2.6.1
python3-openid==3.0.10
pytz==2016.7
six==1.10.0
SQLAlchemy==1.1.1
sqlalchemy-migrate==0.10.0
sqlparse==0.2.1
Tempita==0.5.2
Werkzeug==0.11.11
Whoosh==2.7.4
WTForms==2.1
